Separate SCL functions to separate namespace
SCL - Series common line.
Separate to allow for other connection types in future iterations.
