+++ library/vrfblib/CMakeLists.txt
@@ -19,6 +19,7 @@ add_library(vrfblib SHARED
 
     include/private/celleff.hpp
     include/private/shuntcur/shuntcur.hpp
+    include/private/shuntcur/conn_scl.hpp
 
     src/celleff.cpp
     src/shuntcur.cpp
